Scripts and Classic parser and SFDouble fields - some issues passing these things in/out of scripts.

+++ freex3d/src/lib/vrml_parser/CParseLexer.c
@@ -1097,7 +1097,7 @@ BOOL lexer_double(struct VRMLLexer* me, vrmlDoubleT* ret)
 
   if(negExp)
    exp=-exp;
-  *ret*=(float)(pow(10, exp));
+  *ret*=(pow(10, exp));
  }
  LEXER_UNGETINPUT(c)
